refactor: UserService interface / implementation으로 변경
